// src/generated/models/Crf82_auditlogsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf82_auditlogsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf82_auditlogsstatecode = keyof typeof Crf82_auditlogsstatecode;
export const Crf82_auditlogsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf82_auditlogsstatuscode = keyof typeof Crf82_auditlogsstatuscode;

export interface Crf82_auditlogsBase {
  crf82_actionperformed?: string;
  crf82_activitysummary?: string;
  crf82_activitytimestamp?: string;
  crf82_activitytitle?: string;
  crf82_actorname?: string;
  crf82_auditlogid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf82_auditlogsstatecode;
  statuscode?: Crf82_auditlogsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf82_auditlogs extends Crf82_auditlogsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}